Add support for symmetry groups, changed syntax
This patch adds the ability for clustering to
take into account degenerate groups by symmetry.
For example, a monomer-tetramer interaction will
likely involve 4-fold symmetry, which may or may
not be easily removed by --ignore_chain.

This fix involves a change in the contacts syntax to
seg1 seg2 resid1 resid2, marked by ~~CONTACT_SYNTAX~~
in the code. The %02d is necessary to support
many body interactions (>10 chains) and non-HADDOCK input.